+++ server/src-lib/Hasura/GraphQL/Execute/Postgres.hs
@@ -4,20 +4,21 @@ module Hasura.GraphQL.Execute.Postgres () where
 
 import           Hasura.Prelude
 
-import qualified Data.Environment                          as Env
-import qualified Data.HashSet                              as Set
-import qualified Data.Sequence                             as Seq
-import qualified Language.GraphQL.Draft.Syntax             as G
-import qualified Network.HTTP.Client                       as HTTP
-import qualified Network.HTTP.Types                        as HTTP
+import qualified Data.Environment                           as Env
+import qualified Data.HashSet                               as Set
+import qualified Data.Sequence                              as Seq
+import qualified Language.GraphQL.Draft.Syntax              as G
+import qualified Network.HTTP.Client                        as HTTP
+import qualified Network.HTTP.Types                         as HTTP
 
-import qualified Hasura.Backends.Postgres.Execute.Mutation as PGE
-import qualified Hasura.RQL.IR.Delete                      as IR
-import qualified Hasura.RQL.IR.Insert                      as IR
-import qualified Hasura.RQL.IR.Returning                   as IR
-import qualified Hasura.RQL.IR.Select                      as IR
-import qualified Hasura.RQL.IR.Update                      as IR
-import qualified Hasura.Tracing                            as Tracing
+import qualified Hasura.Backends.Postgres.Execute.LiveQuery as PGL
+import qualified Hasura.Backends.Postgres.Execute.Mutation  as PGE
+import qualified Hasura.RQL.IR.Delete                       as IR
+import qualified Hasura.RQL.IR.Insert                       as IR
+import qualified Hasura.RQL.IR.Returning                    as IR
+import qualified Hasura.RQL.IR.Select                       as IR
+import qualified Hasura.RQL.IR.Update                       as IR
+import qualified Hasura.Tracing                             as Tracing
 
 import           Hasura.Backends.Postgres.Connection
 import           Hasura.EncJSON
@@ -25,33 +26,30 @@ import           Hasura.GraphQL.Context
 import           Hasura.GraphQL.Execute.Backend
 import           Hasura.GraphQL.Execute.Common
 import           Hasura.GraphQL.Execute.Insert
+import           Hasura.GraphQL.Execute.LiveQuery.Plan
 import           Hasura.GraphQL.Execute.Prepare
 import           Hasura.GraphQL.Parser
 import           Hasura.RQL.Types
-import           Hasura.Server.Version                     (HasVersion)
+import           Hasura.Server.Version                      (HasVersion)
 import           Hasura.Session
 
 
-instance
-  ( MonadIO tx
-  , MonadTx tx
-  , Tracing.MonadTrace tx
-  ) => BackendExecute 'Postgres tx where
-  type PreparedQuery 'Postgres = PreparedSql
+instance BackendExecute 'Postgres where
+  type PreparedQuery    'Postgres = PreparedSql
+  type MultiplexedQuery 'Postgres = PGL.MultiplexedQuery
+  type ExecutionMonad   'Postgres = Tracing.TraceT (LazyTxT QErr IO)
   getRemoteJoins = concatMap (toList . snd) . maybe [] toList . _psRemoteJoins
 
   mkDBQueryPlan = pgDBQueryPlan
   mkDBMutationPlan = pgDBMutationPlan
+  mkDBSubscriptionPlan = pgDBSubscriptionPlan
 
 
 -- query
 
 pgDBQueryPlan
-  :: forall m tx .
+  :: forall m .
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -61,7 +59,7 @@ pgDBQueryPlan
   -> [G.Directive G.Name]
   -> SourceConfig 'Postgres
   -> QueryDB 'Postgres (UnpreparedValue 'Postgres)
-  -> m (ExecutionStep tx)
+  -> m ExecutionStep
 pgDBQueryPlan env manager reqHeaders userInfo _directives sourceConfig qrf = do
   (preparedQuery, PlanningSt _ _ planVals expectedVariables) <- flip runStateT initPlanningSt $ traverseQueryDB prepareWithPlan qrf
   validateSessionVariables expectedVariables $ _uiSession userInfo
@@ -72,11 +70,8 @@ pgDBQueryPlan env manager reqHeaders userInfo _directives sourceConfig qrf = do
 -- mutation
 
 convertDelete
-  :: forall m tx .
+  :: forall m .
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -84,18 +79,15 @@ convertDelete
   -> PGE.MutationRemoteJoinCtx
   -> IR.AnnDelG 'Postgres (UnpreparedValue 'Postgres)
   -> Bool
-  -> m (tx EncJSON)
+  -> m (Tracing.TraceT (LazyTxT QErr IO) EncJSON)
 convertDelete env usrVars remoteJoinCtx deleteOperation stringifyNum = do
   let (preparedDelete, expectedVariables) = flip runState Set.empty $ IR.traverseAnnDel prepareWithoutPlan deleteOperation
   validateSessionVariables expectedVariables usrVars
   pure $ PGE.execDeleteQuery env stringifyNum (Just remoteJoinCtx) (preparedDelete, Seq.empty)
 
 convertUpdate
-  :: forall m tx .
+  :: forall m.
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -103,7 +95,7 @@ convertUpdate
   -> PGE.MutationRemoteJoinCtx
   -> IR.AnnUpdG 'Postgres (UnpreparedValue 'Postgres)
   -> Bool
-  -> m (tx EncJSON)
+  -> m (Tracing.TraceT (LazyTxT QErr IO) EncJSON)
 convertUpdate env usrVars remoteJoinCtx updateOperation stringifyNum = do
   let (preparedUpdate, expectedVariables) = flip runState Set.empty $ IR.traverseAnnUpd prepareWithoutPlan updateOperation
   if null $ IR.uqp1OpExps updateOperation
@@ -113,11 +105,8 @@ convertUpdate env usrVars remoteJoinCtx updateOperation stringifyNum = do
     pure $ PGE.execUpdateQuery env stringifyNum (Just remoteJoinCtx) (preparedUpdate, Seq.empty)
 
 convertInsert
-  :: forall m tx .
+  :: forall m.
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -125,7 +114,7 @@ convertInsert
   -> PGE.MutationRemoteJoinCtx
   -> IR.AnnInsert 'Postgres (UnpreparedValue 'Postgres)
   -> Bool
-  -> m (tx EncJSON)
+  -> m (Tracing.TraceT (LazyTxT QErr IO) EncJSON)
 convertInsert env usrVars remoteJoinCtx insertOperation stringifyNum = do
   let (preparedInsert, expectedVariables) = flip runState Set.empty $ traverseAnnInsert prepareWithoutPlan insertOperation
   validateSessionVariables expectedVariables usrVars
@@ -134,11 +123,8 @@ convertInsert env usrVars remoteJoinCtx insertOperation stringifyNum = do
 -- | A pared-down version of 'Query.convertQuerySelSet', for use in execution of
 -- special case of SQL function mutations (see 'MDBFunction').
 convertFunction
-  :: forall m tx .
+  :: forall m.
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -148,7 +134,7 @@ convertFunction
   -> JsonAggSelect
   -> IR.AnnSimpleSelG 'Postgres (UnpreparedValue 'Postgres)
   -- ^ VOLATILE function as 'SelectExp'
-  -> m (tx EncJSON)
+  -> m (Tracing.TraceT (LazyTxT QErr IO) EncJSON)
 convertFunction env userInfo manager reqHeaders jsonAggSelect unpreparedQuery = do
   -- Transform the RQL AST into a prepared SQL query
   (preparedQuery, PlanningSt _ _ planVals expectedVariables)
@@ -165,11 +151,8 @@ convertFunction env userInfo manager reqHeaders jsonAggSelect unpreparedQuery = 
         irToRootFieldPlan planVals $ queryResultFn preparedQuery
 
 pgDBMutationPlan
-  :: forall m tx .
+  :: forall m.
      ( MonadError QErr m
-     , MonadIO tx
-     , MonadTx tx
-     , Tracing.MonadTrace tx
      , HasVersion
      )
   => Env.Environment
@@ -179,7 +162,7 @@ pgDBMutationPlan
   -> Bool
   -> SourceConfig 'Postgres
   -> MutationDB 'Postgres (UnpreparedValue 'Postgres)
-  -> m (ExecutionStep tx)
+  -> m ExecutionStep
 pgDBMutationPlan env manager reqHeaders userInfo stringifyNum sourceConfig mrf =
   ExecStepDB sourceConfig Nothing [] <$> case mrf of
     MDBInsert s              -> convertInsert env userSession remoteJoinCtx s stringifyNum
@@ -189,3 +172,37 @@ pgDBMutationPlan env manager reqHeaders userInfo stringifyNum sourceConfig mrf =
   where
     userSession = _uiSession userInfo
     remoteJoinCtx = (manager, reqHeaders, userInfo)
+
+
+-- mutation
+
+pgDBSubscriptionPlan
+  :: forall m
+  . ( MonadError QErr m
+    , MonadIO m
+    )
+  => UserInfo
+  -> SourceConfig 'Postgres
+  -> InsOrdHashMap G.Name (QueryDB 'Postgres (UnpreparedValue 'Postgres))
+  -> m (LiveQueryPlan 'Postgres (MultiplexedQuery 'Postgres))
+pgDBSubscriptionPlan userInfo sourceConfig unpreparedAST = do
+  (preparedAST, PGL.QueryParametersInfo{..}) <- flip runStateT mempty $
+    for unpreparedAST $ traverseQueryDB PGL.resolveMultiplexedValue
+  let multiplexedQuery = PGL.mkMultiplexedQuery preparedAST
+      roleName = _uiRole userInfo
+      parameterizedPlan = ParameterizedLiveQueryPlan roleName multiplexedQuery
+
+  -- We need to ensure that the values provided for variables are correct according to Postgres.
+  -- Without this check an invalid value for a variable for one instance of the subscription will
+  -- take down the entire multiplexed query.
+  validatedQueryVars     <- PGL.validateVariables (_pscExecCtx sourceConfig) _qpiReusableVariableValues
+  validatedSyntheticVars <- PGL.validateVariables (_pscExecCtx sourceConfig) $ toList _qpiSyntheticVariableValues
+
+  -- TODO validatedQueryVars validatedSyntheticVars
+  let cohortVariables = mkCohortVariables
+        _qpiReferencedSessionVariables
+        (_uiSession userInfo)
+        validatedQueryVars
+        validatedSyntheticVars
+
+  pure $ LiveQueryPlan parameterizedPlan sourceConfig cohortVariables