+++ server/src-lib/Hasura/Server/Init.hs
@@ -6,26 +6,26 @@ module Hasura.Server.Init
   , module Hasura.Server.Init.Config
   ) where
 
-import qualified Data.Aeson                          as J
-import qualified Data.Aeson.TH                       as J
-import qualified Data.HashSet                        as Set
-import qualified Data.String                         as DataString
-import qualified Data.Text                           as T
-import qualified Database.PG.Query                   as Q
-import qualified Language.Haskell.TH.Syntax          as TH
-import qualified Text.PrettyPrint.ANSI.Leijen        as PP
+import qualified Data.Aeson                               as J
+import qualified Data.Aeson.TH                            as J
+import qualified Data.HashSet                             as Set
+import qualified Data.String                              as DataString
+import qualified Data.Text                                as T
+import qualified Database.PG.Query                        as Q
+import qualified Language.Haskell.TH.Syntax               as TH
+import qualified Text.PrettyPrint.ANSI.Leijen             as PP
 
-import           Data.FileEmbed                      (embedStringFile)
-import           Data.Time                           (NominalDiffTime)
+import           Data.FileEmbed                           (embedStringFile)
+import           Data.Time                                (NominalDiffTime)
 import           Data.URL.Template
-import           Network.Wai.Handler.Warp            (HostPreference)
-import qualified Network.WebSockets                  as WS
+import           Network.Wai.Handler.Warp                 (HostPreference)
+import qualified Network.WebSockets                       as WS
 import           Options.Applicative
 
-import qualified Hasura.Cache.Bounded                as Cache
-import qualified Hasura.GraphQL.Execute.LiveQuery    as LQ
-import qualified Hasura.GraphQL.Execute.Plan         as E
-import qualified Hasura.Logging                      as L
+import qualified Hasura.Cache.Bounded                     as Cache
+import qualified Hasura.GraphQL.Execute.LiveQuery.Options as LQ
+import qualified Hasura.GraphQL.Execute.Plan              as E
+import qualified Hasura.Logging                           as L
 
 import           Hasura.Backends.Postgres.Connection
 import           Hasura.Prelude
@@ -37,7 +37,7 @@ import           Hasura.Server.Logging
 import           Hasura.Server.Types
 import           Hasura.Server.Utils
 import           Hasura.Session
-import           Network.URI                         (parseURI)
+import           Network.URI                              (parseURI)
 
 getDbId :: Q.TxE QErr Text
 getDbId =